+++ core/common/timeprofile.h
@@ -94,7 +94,7 @@ static inline void timeprofile_display(timeprofile_t *tp)
 {
     int64_t lastutime = tp->utime;
 
-    for (int i = 0; i < zarray_size(tp->stamps); i++) {
+    for (size_t i = 0; i < zarray_size(tp->stamps); i++) {
         struct timeprofile_entry *stamp;
 
         zarray_get_volatile(tp->stamps, i, &stamp);
@@ -103,7 +103,7 @@ static inline void timeprofile_display(timeprofile_t *tp)
 
         double parttime = (stamp->utime - lastutime)/1000000.0;
 
-        printf("%2d %32s %15f ms %15f ms\n", i, stamp->name, parttime*1000, cumtime*1000);
+        printf("%2ld %32s %15f ms %15f ms\n", i, stamp->name, parttime*1000, cumtime*1000);
 
         lastutime = stamp->utime;
     }
